+++ code/data-migration/account-v0.preview.en.rs
@@ -1,0 +1,11 @@
+#[derive(BorshDeserialize, BorshSerialize, Debug, Default, PartialEq)]
+pub struct AccountContentCurrent {
+    pub somevalue: u64,
+}
+
+#[derive(BorshDeserialize, BorshSerialize, Debug, Default, PartialEq)]
+pub struct ProgramAccountState {
+    is_initialized: bool,
+    data_version: u8,
+    account_data: AccountContentCurrent,
+}
