+++ code/data-migration/account-v1.en.rs
@@ -3,17 +3,25 @@
 use arrayref::{array_ref, array_refs};
 use borsh::{BorshDeserialize, BorshSerialize};
 use solana_program::{
+    borsh::try_from_slice_unchecked,
     msg,
     program_error::ProgramError,
     program_pack::{IsInitialized, Pack, Sealed},
 };
 use std::{io::BufWriter, mem};
 
-/// Currently using state. If version changes occur, this
-/// should be copied to another serializable backlevel one
-/// before adding new fields here
+/// Current state (DATA_VERSION 1). If version changes occur, this
+/// should be copied to another (see AccountContentOld below)
+/// We've added a new field: 'somestring'
 #[derive(BorshDeserialize, BorshSerialize, Debug, Default, PartialEq)]
 pub struct AccountContentCurrent {
+    pub somevalue: u64,
+    pub somestring: String,
+}
+
+/// Old content state (DATA_VERSION 0).
+#[derive(BorshDeserialize, BorshSerialize, Debug, Default, PartialEq)]
+pub struct AccountContentOld {
     pub somevalue: u64,
 }
 
@@ -49,26 +57,27 @@ impl ProgramAccountState {
 }
 
 /// Declaration of the current data version.
-pub const DATA_VERSION: u8 = 0;
+const DATA_VERSION: u8 = 1; // Adding string to content
+                            // Previous const DATA_VERSION: u8 = 0;
+
 /// Account allocated size
-pub const ACCOUNT_ALLOCATION_SIZE: usize = 1024;
+const ACCOUNT_ALLOCATION_SIZE: usize = 1024;
 /// Initialized flag is 1st byte of data block
 const IS_INITIALIZED: usize = 1;
 /// Data version (current) is 2nd byte of data block
 const DATA_VERSION_ID: usize = 1;
 
 /// Previous content data size (before changing this is equal to current)
-pub const PREVIOUS_VERSION_DATA_SIZE: usize = mem::size_of::<AccountContentCurrent>();
+const PREVIOUS_VERSION_DATA_SIZE: usize = mem::size_of::<AccountContentOld>();
 /// Total space occupied by previous account data
-pub const PREVIOUS_ACCOUNT_SPACE: usize =
-    IS_INITIALIZED + DATA_VERSION_ID + PREVIOUS_VERSION_DATA_SIZE;
+const PREVIOUS_ACCOUNT_SPACE: usize = IS_INITIALIZED + DATA_VERSION_ID + PREVIOUS_VERSION_DATA_SIZE;
 
 /// Current content data size
-pub const CURRENT_VERSION_DATA_SIZE: usize = mem::size_of::<AccountContentCurrent>();
+const CURRENT_VERSION_DATA_SIZE: usize = mem::size_of::<AccountContentCurrent>();
 /// Total usage for data only
-pub const CURRENT_USED_SIZE: usize = IS_INITIALIZED + DATA_VERSION_ID + CURRENT_VERSION_DATA_SIZE;
+const CURRENT_USED_SIZE: usize = IS_INITIALIZED + DATA_VERSION_ID + CURRENT_VERSION_DATA_SIZE;
 /// How much of 1024 is used
-pub const CURRENT_UNUSED_SIZE: usize = ACCOUNT_ALLOCATION_SIZE - CURRENT_USED_SIZE;
+const CURRENT_UNUSED_SIZE: usize = ACCOUNT_ALLOCATION_SIZE - CURRENT_USED_SIZE;
 /// Current space used by header (initialized, data version and Content)
 pub const ACCOUNT_STATE_SPACE: usize = CURRENT_USED_SIZE + CURRENT_UNUSED_SIZE;
 
@@ -76,20 +85,25 @@ pub const ACCOUNT_STATE_SPACE: usize = CURRENT_USED_SIZE + CURRENT_UNUSED_SIZE;
 /// to current state of data
 fn conversion_logic(src: &[u8]) -> Result<ProgramAccountState, ProgramError> {
     let past = array_ref![src, 0, PREVIOUS_ACCOUNT_SPACE];
-    let (initialized, _, _account_space) = array_refs![
+    let (initialized, _, account_space) = array_refs![
         past,
         IS_INITIALIZED,
         DATA_VERSION_ID,
         PREVIOUS_VERSION_DATA_SIZE
     ];
-    // Logic to uplift from previous version
+    // Logic to upgrade from previous version
     // GOES HERE
+    let old = try_from_slice_unchecked::<AccountContentOld>(account_space).unwrap();
+    // Default sets 'somevalue' to 0 and somestring to default ""
+    let mut new_content = AccountContentCurrent::default();
+    // We copy the existing 'somevalue', the program instructions will read/update 'somestring' without fail
+    new_content.somevalue = old.somevalue;
 
     // Give back
     Ok(ProgramAccountState {
         is_initialized: initialized[0] != 0u8,
         data_version: DATA_VERSION,
-        account_data: AccountContentCurrent::default(),
+        account_data: new_content,
     })
 }
 impl Sealed for ProgramAccountState {}
@@ -116,11 +130,8 @@ impl Pack for ProgramAccountState {
         if initialized {
             // Version check
             if src[1] == DATA_VERSION {
-                msg!("Processing consistent data");
-                Ok(
-                    ProgramAccountState::try_from_slice(array_ref![src, 0, CURRENT_USED_SIZE])
-                        .unwrap(),
-                )
+                msg!("Processing consistent version data");
+                Ok(try_from_slice_unchecked::<ProgramAccountState>(src).unwrap())
             } else {
                 msg!("Processing backlevel data");
                 conversion_logic(src)
